modularize gameplay screen by extracting logic and event handlers into custom hooks
